Split routeplaner into files
