+++ frontend/src/components/RoutePlanner/RoutePlanner.tsx
@@ -17,25 +17,17 @@ import { formatISO, isSameMinute, parseISO } from "date-fns";
 import { closeCircleOutline } from "ionicons/icons";
 import { useState } from "react";
 import {
-  FootpathLeg,
-  Journey,
-  LegOriginLocationTypeEnum,
-  Location,
-  TransportationLeg,
-  TransportationLegTypeEnum
-} from "../api";
-import { useCurrentTime } from "../hooks/useCurrentTime";
-import { useCustomDepartureTimeUrlParamOrCurrentTime } from "../hooks/useCustomDepartureTimeOrCurrentTime";
-import { useJourneyQuery } from "../hooks/useJourneyQuery";
-import { useLocationByIdOrNull } from "../hooks/useLocationByIdOrNull";
-import { useStateParams } from "../hooks/useStateParams";
-import { IJourney } from "../interfaces/IJourney.interface";
-import { IJourneyStep } from "../interfaces/IJourneyStep.interface";
-import FavoritesPage from "../pages/FavoritesPage";
-import { CreateFavoriteRoute, CreateFavoriteTrip, useFavoriteRoutes, useFavoriteTrips } from "../services/favorites/FavoritesContext";
-import JourneyListComponent from "./JourneyListComponent";
-import { LocationSearchInput } from "./LocationSearch/LocationSearchInput";
+  Location
+} from "../../api";
+import { useCurrentTime } from "../../hooks/useCurrentTime";
+import { useCustomDepartureTimeUrlParamOrCurrentTime } from "../../hooks/useCustomDepartureTimeOrCurrentTime";
+import { useLocationByIdOrNull } from "../../hooks/useLocationByIdOrNull";
+import { useStateParams } from "../../hooks/useStateParams";
+import FavoritesPage from "../../pages/FavoritesPage";
+import { CreateFavoriteRoute, CreateFavoriteTrip, useFavoriteRoutes, useFavoriteTrips } from "../../services/favorites/FavoritesContext";
+import { LocationSearchInput } from "../LocationSearch/LocationSearchInput";
 import "./RoutePlanner.css";
+import { TripOptionsDisplay } from "./TripOptionsDisplay";
 
 export const DEPARTURE_TIME_NOW_PARAM = "now";
 
@@ -198,7 +190,7 @@ const RoutePlanner = ({ setSelectedOriginLocation, setSelectedDestinationLocatio
           departure={departureTime}
         />
       }
-      <IonModal id="favorite-dialogue" isOpen={isFavoriteDialogueOpen} onDidDismiss={() => setIsFavoritesDialogueOpen(false)}>
+      <IonModal isOpen={isFavoriteDialogueOpen} onDidDismiss={() => setIsFavoritesDialogueOpen(false)}>
         <IonContent>
           <IonToolbar>
             <IonTitle>Add to favorites</IonTitle>
@@ -243,77 +235,3 @@ const RoutePlanner = ({ setSelectedOriginLocation, setSelectedDestinationLocatio
 };
 
 export default RoutePlanner;
-
-export function TripOptionsDisplay(props: {
-  origin: Location,
-  destination: Location,
-  departure: Date
-}): JSX.Element {
-  const { origin, destination, departure } = props;
-
-  // TODO Add user input if datetime should be interpreted as arrival time.
-  const result = useJourneyQuery(origin, destination, departure, false);
-
-  const iJourneys: false | IJourney[] = result.type === "success"
-    && result.journeyResults
-      .map((journey): IJourney => {
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        const legs: (TransportationLeg | FootpathLeg)[] = journey.legs;
-
-        const lastLeg = legs[legs.length - 1];
-        const firstLeg = legs[0];
-
-        return {
-          startStation: firstLeg.origin.name,
-          startTime: firstLeg.origin.departureTimeEstimated,
-          arrivalStation: lastLeg.destination.name,
-          arrivalTime: lastLeg.destination.arrivalTimeEstimated,
-          stops: legs.map((leg): IJourneyStep => ({
-            arrivalTime: leg.destination.arrivalTimeEstimated,
-            startTime: leg.origin.departureTimeEstimated,
-            stationName: leg.origin.name,
-            track: leg.origin.type === LegOriginLocationTypeEnum.Platform
-              ? leg.origin.details.shortName
-              : "",
-            stopName: leg.destination.name,
-            travelDurationInMinutes: leg.details.duration / 60
-          })),
-          travelDurationInMinutes: legs.reduce((acc, leg) => acc + leg.details.duration, 0) / 60
-        };
-      });
-
-  return (
-    <>
-      {result.type === "error" && <div>Error: {result.error.message}</div>}
-      {result.type === "pending" && <div>Searching...</div>}
-      {iJourneys &&
-        <JourneyListComponent journeys={iJourneys} />
-      }
-    </>
-  );
-}
-
-export function RenderTrip(props: { journey: Journey }): JSX.Element {
-  const { journey } = props;
-
-  return (
-    <IonItem>
-      <IonLabel>
-        <ol>
-          {
-            journey.legs.map((leg, idx) => (
-              <li key={idx}>
-                {
-                  leg.type === TransportationLegTypeEnum.Transportation
-                    ? (leg as TransportationLeg).transportation.name
-                    : "Footpath"
-                }
-                {leg.details.duration}
-              </li>
-            ))
-          }
-        </ol>
-      </IonLabel>
-    </IonItem>
-  );
-}
