+++ frontend/src/pages/JourneyPage.tsx
@@ -2,7 +2,7 @@ import { IonButtons, IonContent, IonHeader, IonImg, IonMenuButton, IonPage, IonT
 import React, { useState } from "react";
 import logo from "../../public/images/train_image.png";
 import { Location } from "../api";
-import RoutePlanner from "../components/RoutePlanner";
+import RoutePlanner from "../components/RoutePlanner/RoutePlanner";
 import LeafletMapContainer from "../components/map/LeafletMapContainer";
 import "./JourneyPage.css";
